# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MixPlay API",
        "description": "API da aplicação MixPlay, para pegar os dados.",
        "contact": {
            "email": "rodrigo.pinheiroa@hotmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Default Server"
        }
    ],
    "paths": {
...
        "/tmdb/search/{query}": {
            "get": {
                "tags": [
                    "TMDB"
                ],
                "summary": "Buscar por filmes ou séries",
                "operationId": "tmdbSearch",
                "parameters": [
                    {
                        "name": "query",
                        "in": "path",
                        "description": "Texto a ser pesquisado",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Página",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/TMDB\\Search Collection"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#components/responses/unauthorized"
                    }
                }
            }
        },
        "/tmdb/movie/{id}": {
            "get": {
                "tags": [
                    "TMDB"
                ],
                "summary": "Detalhar filme por id",
                "operationId": "tmdbMovieDetail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "TMDB ID do filme",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/TMDB\\Movie"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#components/responses/unauthorized"
                    }
                }
            }
        },
        "/tmdb/tvshow/{id}": {
            "get": {
                "tags": [
                    "TMDB"
                ],
                "summary": "Detalhar série por id",
                "operationId": "tmdbSerieDetail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "TMDB ID da série",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/TMDB\\TvShow"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#components/responses/unauthorized"
                    }
                }
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}